+++ MyStack/MyStack.cpp
@@ -8,9 +8,9 @@ using namespace std;
 int main()
 {
 	stack<int> a;
-	for (size_t i = 0; i < 10; i++)
+	for (int i = 0; i < 10; i++)
 		a.push(i);
-	for (size_t i = 0; i < 10; i++)
+	for (int i = 0; i < 10; i++)
 	{
 		cout << a.top() << endl;
 		a.pop();
@@ -20,7 +20,7 @@ int main()
 
 	MyStack<int> b;
 
-	for (size_t i = 0; i < 10; i++)
+	for (int i = 0; i < 10; i++)
 		b.Push(i);
 
 	for (size_t i = 0; i < 10; i++)
